% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_demo_variables}
\alias{census_demo_variables}
\title{Census Demographic Variables}
\format{
\subsection{\code{census_demo_variables}}{

A data frame with 68 rows and 2 columns:
\describe{
\item{var_code}{Census Variable Code}
\item{var_name}{Census Variable Name}
}
}
}
\usage{
census_demo_variables
}
\description{
Census Demographic Variables
}
\keyword{datasets}
