% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{councilverse_packages}
\alias{councilverse_packages}
\title{List all packages imported by councilverse}
\usage{
councilverse_packages()
}
\description{
List all packages imported by councilverse
}
\examples{
councilverse_packages()
}
