% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_html_helper_functions.R
\name{encodeGraphic}
\alias{encodeGraphic}
\title{Graphics to raw HTML}
\usage{
encodeGraphic(g, style = "style=\\"width:150px;height:100px\\"", ...)
}
\arguments{
\item{g}{plot or image object.}

\item{style}{html style, default is set to width of 150px and height of 100px}

\item{...}{}
}
\value{
Raw HTML for the image
}
\description{
Helper function to convert a plot or image to raw HTML
}
