% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_name_generator_function.R
\name{file_name_generator}
\alias{file_name_generator}
\title{Helper function for creating file names using accepted guidelines}
\usage{
file_name_generator(
  order = NULL,
  description = NULL,
  time_granularity = NULL,
  disaggregation_categories = NULL,
  date_year = NULL,
  file_extension = NULL,
  ...
)
}
\arguments{
\item{order}{(string). The sequential number in \code{##} format (if applicable) when the file is run}

\item{description}{(string). The dateset name, source, or location}

\item{time_granularity}{(string). Time granularity (e.g. 'hourly', 'daily', 'minutely', 'yearly', etc.)}

\item{disaggregation_categories}{(string). Grouping categorizer; accepts vectors (e.g. 'age','cd',c('age','cd'))}

\item{date_year}{(numeric). Date or year}

\item{file_extension}{(string). File extension (e.g. ".csv")}

\item{...}{Add anything else to the back of the file name before file extension}
}
\value{
A string with the file output name
}
\description{
\enumerate{
\item Order file should be run (00-10)
\item Dataset Name, Source, Location
\item Time granularity (hourly, daily, minutely, yearly, etc.)
\item Grouping categorizer ex ‘by-age’, ‘by-cd’, ‘by-cd-age’
\item Date or Year
}
}
\examples{
\dontrun{
# All fields used
file_name_generator(order = "01", description = "acs_poverty", time_granularity = "daily", disaggregation_categories = c("cd", "race"), date_year = 2018, file_extension = ".R",... = "example")
# Shorter example
file_name_generator(description = "puma_internet", date_year = 2022)
}
}
