% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_estimates.R
\name{get_ACS_demographic_estimates}
\alias{get_ACS_demographic_estimates}
\title{get_demo_estimates}
\usage{
get_ACS_demographic_estimates(
  var_code_list,
  geo,
  polygons = FALSE,
  download = FALSE
)
}
\value{
printing
}
\description{
python
}
