% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_estimates.R
\name{get_BBL_estimates}
\alias{get_BBL_estimates}
\title{Get the Building Block Lot (BBL) level estimates for all available census variables}
\usage{
get_BBL_estimates()
}
\value{
DataFrame of BBL level estimates
}
\description{
Get the Building Block Lot (BBL) level estimates for all available census variables
}
\examples{
\dontrun{
library(councilverse)
get_BBL_estimates()
}
}
