% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_variables_function.R
\name{get_census_variables}
\alias{get_census_variables}
\title{Function to retrieve the available census demographic variables}
\usage{
get_census_variables()
}
\value{
data frame of available variables
}
\description{
Function to retrieve the available census demographic variables
}
