% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_html_helper_functions.R
\name{gt_table}
\alias{gt_table}
\title{gt table to raw HTML helper function}
\usage{
gt_table(x, currency_col = NA, percent_col = NA, extra_formats = NULL)
}
\arguments{
\item{x}{gt table}

\item{currency_col}{add currency formatting to all columns specified}

\item{percent_col}{add percent formatting to all columns specified}

\item{extra_formats}{add any additional formatting}
}
\value{
Raw HTML for the gt table
}
\description{
Helper function to convert gt tables to raw HTML for side-by-side comparisons
}
