% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_helper_functions.R
\name{unzip_sf}
\alias{unzip_sf}
\title{Helper function for loading in shapefiles that come zipped}
\usage{
unzip_sf(zip_url)
}
\arguments{
\item{zip_url}{The url that the file you need to download and unzip is at}
}
\value{
The temporary path the file is stored in, ready to pass to st_read() or other
}
\description{
Helper function for loading in shapefiles that come zipped
}

\examples{
\dontrun{
url <- "https://www1.nyc.gov/assets/planning/download/zip/data-maps/open-data/nyct2020_22a.zip"
# unzip the zip file
zip <- unzip_sf(url)
# assign to variable with read_csv or other read functions
sf <- sf::read_sf(zip)
}
}
