% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_estimates.R
\name{view_ACS_variables}
\alias{view_ACS_variables}
\title{View available census variables and their descriptions}
\usage{
view_ACS_variables(as_df = TRUE)
}
\arguments{
\item{as_df}{boolean. \code{as_df = TRUE} to return a dataframe. \code{as_df = FALSE} to just print. default \code{as_df = TRUE}}
}
\value{
Either a DataFrame or printed census variable codes and their description
}
\description{
View available census variables and their descriptions
}
\examples{
\dontrun{
library(councilverse)
view_ACS_variables()
view_ACS_variables(as_df = FALSE)
}
}
