% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bbl_estimates_function.R
\name{get_bbl_estimates}
\alias{get_bbl_estimates}
\title{Function to retrieve PLUTO bbl population estimates}
\usage{
get_bbl_estimates()
}
\value{
tibble with PLUTO bbl population estimates
}
\description{
Function to retrieve PLUTO bbl population estimates
}
