% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_estimates_function.R
\name{get_geo_estimates}
\alias{get_geo_estimates}
\title{Function to retrieve acs estimates from csvs at different geographies}
\usage{
get_geo_estimates(name = NULL)
}
\arguments{
\item{name}{Name of the csv. If NULL, returns a list of possible geography csvs.}
}
\value{
sf for the specified geography, or list of geographies if none is specified
}
\description{
Function to retrieve acs estimates from csvs at different geographies
}
